Redefined the uses of GriefAction classes
The original "ImmutableGriefAction" class was used as a stagnant version
of the Grief Action. This was so the configuration file upon load could
establish which actions were to be flagged. Then, a specific Grief
Action can be called when such an event occured, and a griefer could be
"assigned" to this action. I renamed these classes and redefined their
construction to better forge this relationship. Copying no longer occurs
and a "GriefInstance" now houses information such as griefer and entity
snapshots.