Reformatted the way the GriefAlert data is held
Before, there were three static tables which held information about all
the GriefActions. These all came from the text file which was read when
the plugin loaded. There were numerous, repetitive methods to manage all
these static tables, which became confusing and adding to tables became
convoluted. I combined all of the tables into a single class, which can
be initialized and GriefActions can be put in using the same parameters
as a generic Table, but with the additional GriefType parameter to know
into which table to load it.