Changed the parameters of some methods in Main class: GriefAlert
The main class for the plugin (GriefAlert) had many static methods that
many other classes used to get information such as configuration files
and such. This didn't ideal because the plugin itself is run as one
instance of the GriefAlert class, so all other classes should be
instances held within that one instance. I removed many of the static
modifiers and reconstructed classes to house the main instance of
GriefAlert. Then, all other instance methods of other classes, like
AlertTracker, is found within the "plugin" field of each class.