Changed names in AlertTracker and Indexing at 0.
I noticed that the first id was changed to be 1 instead of 0 within the
id numbers themselves. I don't consider this to be good practice because
arrays index by 0 everywhere, so this is bound to cause confusion and
complication. The index now starts at 0 again, but with the intention
that player interaction with this array will result in a suitable
number. In other words, the index number will appear one larger than it
actually is to players.