Reformatting AlertTracker -> AlertManager, upgrading utilities
Renaming methods in AlertTracker (and the class itself) makes it easier
to see how alerts are handled by the server. I'm continuing to rename
and document these processes as well as reorganize some generally useful
utilities. An example is a new method in the General class of the tools
package which can correct the indefinite articles ("a" and "an")
appropriately throughout a string.