Javadocs mainly in GriefLogger.java and Re-document Sign Logging
When I first read through the plugin, it seemed like signs just had a
different listener altogether so I thought it was just about griefing
existing signs. I updated the names of the methods and documentation so
it's clear that it's talking about *editing* signs (new in later
versions of MC).